+++ lib/flutter_dropdown_button.dart
@@ -7,4 +7,5 @@ export 'src/config/text_dropdown_config.dart';
 export 'src/theme/dropdown_scroll_theme.dart';
 export 'src/theme/dropdown_style_theme.dart';
 export 'src/theme/dropdown_theme.dart';
+export 'src/theme/search_field_theme.dart';
 export 'src/theme/tooltip_theme.dart';